refactor: 공통 컬럼(createdAt, updatedAt, isDeleted) -> BaseEntity로 분리
