+++ src/main/kotlin/com/ssak3/timeattack/common/domain/BaseEntity.kt
@@ -1,0 +1,24 @@
+package com.ssak3.timeattack.common.domain
+
+import jakarta.persistence.Column
+import jakarta.persistence.MappedSuperclass
+import jakarta.persistence.PreUpdate
+import java.time.LocalDateTime
+
+@MappedSuperclass
+abstract class BaseEntity {
+
+    @Column(name = "created_at", updatable = false)
+    val createdAt: LocalDateTime = LocalDateTime.now()
+
+    @Column(name = "updated_at")
+    var updatedAt: LocalDateTime = LocalDateTime.now()
+
+    @Column(name = "is_deleted")
+    var isDeleted: Boolean = false
+
+    @PreUpdate
+    fun preUpdate() {
+        updatedAt = LocalDateTime.now()
+    }
+}
